mac80211: minstrel: fix variance overestimation / fluctuation
The variance / standard deviation is artificially inflated for rates
where only sample attempts are counted. With only few packet attempts,
the difference between the current probabilty and the ewma value tends
to be very big because of the lack of precision in the current value.

Whenever not enough data was gathered, pass in the previous average
instead of the current probability. This leads to the standard deviation
value slowly decreasing for sampled rates.

Signed-off-by: Felix Fietkau <nbd@openwrt.org>
Signed-off-by: Thomas Huehn <thomas@net.t-labs.tu-berlin.de>