mac80211: change rate downgrade behavior in Minstrel_HT
This changes the behavior of the new rate downgrade approach in
cases where stream errors happen. In case if 50% stream errors we
try now to switch (I) from LGI to SGI, (II) try to decrease the #
of spatial streams and (III) try to reduce channel bandwidth.
In addition there are some code cleanups.

Signed-off-by: Thomas Huehn <thomas@net.t-labs.tu-berlin.de>