ath9k: Fix overflow in ath_buf txpower field
The txpower in ath_buf must not be larger than 63 because this is what 63
because the HW is expecting only a 6 bit value. But it was seen that the
code still tries to write something like 72 to this structure (which
results in only 4 dBm after truncation to 6 bit and conversion from
twicepower to dBm.

The used txpower (in dBm) must therefore not be larger than
MAX_RATE_POWER / 2 to avoid that they overflow after the conversion from
dBm to twicepower.

Signed-off-by: Sven Eckelmann <sven@narfation.org>