mac80211: Fix retrieval of txpower for interface
The code only supported retrieval of the txpower via
sdata->vif.bss_conf.txpower. But this would break when the
local->ops->get_txpower function pointer has to be used or when channel
context were not enabled.

This missing functionality broke for example the txpower on WDS APs because
the retrieved power was always 0.

Signed-off-by: Sven Eckelmann <sven@narfation.org>