mac80211: minstrel(_ht): Apply fixed_txpower also for sampling packets
We must not use suddenly a higher txpower for sampling packets when the
peer fixed txpower is set to a lower txpower. Otherwise the remote rx might
get saturated and can no longer receive the packets we used for sampling.
As result, the sampling fails not because of the rate but of a suddenly
changing txpower.

Signed-off-by: Sven Eckelmann <sven@narfation.org>